PHP 8.3 compatibiliteit en Wheel map verplaatst naar public
1. Precision Loss fix (SymbolGlyph.php)
   - getSignGlyph() parameter: int → float
   - Voorkomt deprecated warning in PHP 8.1+

2. FFI aanroepen gemoderniseerd (SwissEphemeris.php, HouseCalculator.php)
   - \FFI::new() → self::$ffi->new()
   - \FFI::string() → self::$ffi->string()
   - Vereist voor PHP 8.3 compatibiliteit

3. Wheel map verplaatst naar public/
   - Verplaatst van src/Wheel/ naar public/Wheel/
   - Directe browser toegang tot generate_image.php
   - URL aangepast in index.php

4. Debug flags verwijderd (index.php)
   - display_errors verwijderd voor productie

Fixes productie issues op PHP 8.3 server.

+++ src/Glyph/SymbolGlyph.php
@@ -198,7 +198,7 @@ class SymbolGlyph
         return self::$planets[$code]['code'];
     }
 
-    public static function getSignGlyph(int $longitude): string
+    public static function getSignGlyph(float $longitude): string
     {
         $signIndex = (int) floor($longitude / 30);
         $signIndex = max(0, min(11, $signIndex));
